# src/main/java/com/service/management/servicemanagement/model/ActivateServiceId.java
/*
 * Service Management API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package com.service.management.servicemanagement.model;

import java.util.Objects;

import com.google.gson.annotations.SerializedName;
import io.swagger.annotations.ApiModelProperty;
import javax.validation.constraints.NotEmpty;
/**
 * ActivateServiceId
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2022-08-17T07:41:31.670Z")
public class ActivateServiceId {

    @NotEmpty(message = "Please provide a valid customer Id")
    @SerializedName("customerId")
    private String customerId = null;

    @NotEmpty(message = "Please provide a valid service number")
    @SerializedName("serviceId")
    private String serviceId = null;

    public ActivateServiceId customerId(String customerId) {
        this.customerId = customerId;
        return this;
    }

    /**
     * Unique Customer identifier
     * @return customerId
     **/
    @ApiModelProperty(value = "Unique Customer identifier")
    public String getCustomerId() {
        return customerId;
    }

    public void setCustomerId(String customerId) {
        this.customerId = customerId;
    }

    public ActivateServiceId serviceId(String serviceId) {
        this.serviceId = serviceId;
        return this;
    }

    /**
     * Unique Service number.
     * @return serviceId
     **/
    @ApiModelProperty(value = "Unique Service number.")
    public String getServiceId() {
        return serviceId;
    }

    public void setServiceId(String serviceId) {
        this.serviceId = serviceId;
    }


    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        ActivateServiceId activateServiceId = (ActivateServiceId) o;
        return Objects.equals(this.customerId, activateServiceId.customerId) &&
                Objects.equals(this.serviceId, activateServiceId.serviceId);
    }

    @Override
    public int hashCode() {
        return Objects.hash(customerId, serviceId);
    }


    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();
        sb.append("class ActivateServiceId {\n");

        sb.append("    customerId: ").append(toIndentedString(customerId)).append("\n");
        sb.append("    serviceId: ").append(toIndentedString(serviceId)).append("\n");
        sb.append("}");
        return sb.toString();
    }

    /**
     * Convert the given object to string with each line indented by 4 spaces
     * (except the first line).
     */
    private String toIndentedString(Object o) {
        if (o == null) {
            return "null";
        }
        return o.toString().replace("\n", "\n    ");
    }

}